+++ examples/interface.rs
@@ -35,12 +35,12 @@ pub struct Context;
 impl juniper::Context for Context {}
 
 pub struct Article {
-    id: Id,
+    id: ID,
     text: String,
 }
 
 impl ArticleFields for Article {
-    fn field_id(&self, executor: &Executor<'_, Context>) -> FieldResult<&Id> {
+    fn field_id(&self, executor: &Executor<'_, Context>) -> FieldResult<&ID> {
         unimplemented!()
     }
 
@@ -50,12 +50,12 @@ impl ArticleFields for Article {
 }
 
 pub struct Tweet {
-    id: Id,
+    id: ID,
     text: String,
 }
 
 impl TweetFields for Tweet {
-    fn field_id(&self, executor: &Executor<'_, Context>) -> FieldResult<&Id> {
+    fn field_id(&self, executor: &Executor<'_, Context>) -> FieldResult<&ID> {
         unimplemented!()
     }
 
@@ -74,11 +74,11 @@ impl QueryFields for Query {
         query: String,
     ) -> FieldResult<Vec<SearchResult>> {
         let article: Article = Article {
-            id: Id::new("1"),
+            id: ID::new("1"),
             text: "Business".to_string(),
         };
         let tweet: Tweet = Tweet {
-            id: Id::new("2"),
+            id: ID::new("2"),
             text: "1 weird tip".to_string(),
         };