+++ src/walk_ast/find_special_scalar_types.rs
@@ -4,7 +4,6 @@ use graphql_parser::schema::*;
 pub struct SpecialScalarTypesList {
     date_defined: bool,
     date_time_defined: bool,
-    id_scalar_used: bool,
 }
 
 impl SpecialScalarTypesList {
@@ -15,10 +14,6 @@ impl SpecialScalarTypesList {
     pub fn date_time_defined(&self) -> bool {
         self.date_time_defined
     }
-
-    pub fn id_scalar_used(&self) -> bool {
-        self.id_scalar_used
-    }
 }
 
 pub fn find_special_scalar_types(doc: &Document) -> SpecialScalarTypesList {
@@ -28,7 +23,6 @@ pub fn find_special_scalar_types(doc: &Document) -> SpecialScalarTypesList {
     let mut out = SpecialScalarTypesList {
         date_defined: false,
         date_time_defined: false,
-        id_scalar_used: false,
     };
 
     for def in &doc.definitions {
@@ -40,27 +34,6 @@ pub fn find_special_scalar_types(doc: &Document) -> SpecialScalarTypesList {
                     _ => {}
                 },
 
-                Object(obj) => {
-                    for field in &obj.fields {
-                        if is_id_type(&field.field_type) {
-                            out.id_scalar_used = true
-                        }
-
-                        for arg in &field.arguments {
-                            if is_id_type(&arg.value_type) {
-                                out.id_scalar_used = true
-                            }
-                        }
-                    }
-                }
-                InputObject(obj) => {
-                    for field in &obj.fields {
-                        if is_id_type(&field.value_type) {
-                            out.id_scalar_used = true
-                        }
-                    }
-                }
-
                 _ => {}
             },
             _ => {}
@@ -68,14 +41,4 @@ pub fn find_special_scalar_types(doc: &Document) -> SpecialScalarTypesList {
     }
 
     out
-}
-
-fn is_id_type(r#type: &Type) -> bool {
-    use graphql_parser::query::Type::*;
-
-    match r#type {
-        NamedType(name) => name == "ID",
-        ListType(inner) => is_id_type(&inner),
-        NonNullType(inner) => is_id_type(&inner),
-    }
 }